remove spurious casts and parentheses

+++ entrylist.c
@@ -17,7 +17,7 @@
 // random number
 uint32_t irand(uint32_t b, uint32_t e) {
   double r = e - b + 1;
-  return b + (uint32_t) (r * rand() / (RAND_MAX + 1.0));
+  return b + r * rand() / (RAND_MAX + 1.0);
 }
 
 // List functions
